split preset model
