+++ voicevox_engine/preset/__init__.py
@@ -1,1 +1,2 @@
+from .Preset import Preset  # noqa: F401
 from .PresetLoader import PresetLoader  # noqa: F401
